Refactor logic to enumerate cores, and update options
