+++ winml/lib/Api/LearningModelDevice.cpp
@@ -7,6 +7,7 @@
 #include <D3d11_4.h>
 #include <d3d11on12.h>
 #include "D3DDeviceCache.h"
+#include "HardwareCoreEnumerator.h"
 
 #include "ConverterResourceStore.h"
 
@@ -129,79 +130,10 @@ LearningModelDevice::CacheThreadPool(_winml::IThreading* thread_pool) {
   return S_OK;
 }
 
-struct LogicalProcessorInformation {
-  std::unique_ptr<char[]> Buffer;
-  size_t Length;
-};
-
-static LogicalProcessorInformation GetLogicalProcessorInfos(LOGICAL_PROCESSOR_RELATIONSHIP relationship) {
-  DWORD length = 0;
-  DWORD rc = GetLogicalProcessorInformationEx(relationship, nullptr, &length);
-
-  assert(rc == FALSE);
-
-  auto processorInformationBytes = std::make_unique<char[]>(length);
-
-  rc = GetLogicalProcessorInformationEx(
-    relationship, (PSYSTEM_LOGICAL_PROCESSOR_INFORMATION_EX)processorInformationBytes.get(), &length
-  );
-
-  assert(rc == TRUE);
-
-  return {std::move(processorInformationBytes), length};
-}
-
-long long GetNumberOfSoCDieAtoms() {
-  // while (Size > (ULONG)FIELD_OFFSET(SYSTEM_LOGICAL_PROCESSOR_INFORMATION_EX, Processor)) {
-  DWORD dwLevel2GroupMask = 0;
-  DWORD dwLevel3GroupMask = 0;
-  DWORD dwSoCGroupMask = 0;
-
-  auto logicalProcessorInformation = GetLogicalProcessorInfos(RelationAll);
-  auto processorInformation = (PSYSTEM_LOGICAL_PROCESSOR_INFORMATION_EX)logicalProcessorInformation.Buffer.get();
-
-  size_t read = 0;
-  while (read <= logicalProcessorInformation.Length) {
-    switch (processorInformation->Relationship) {
-      case RelationCache:
-        if (processorInformation->Cache.Level == 2) {
-          dwLevel2GroupMask |= processorInformation->Cache.GroupMask.Mask;
-        } else if (processorInformation->Cache.Level == 3) {
-          dwLevel3GroupMask |= processorInformation->Cache.GroupMask.Mask;
-        }
-        break;
-    }
-
-    read += sizeof(SYSTEM_LOGICAL_PROCESSOR_INFORMATION_EX);
-    processorInformation++;
-  }
-
-  dwSoCGroupMask = (dwLevel2GroupMask & ~dwLevel3GroupMask);
-  return __popcnt(dwSoCGroupMask);
-}
-
-long long GetNumberOfCores() {
-  auto logicalProcessorInformation = GetLogicalProcessorInfos(RelationProcessorCore);
-  auto processorInformation = (PSYSTEM_LOGICAL_PROCESSOR_INFORMATION_EX)logicalProcessorInformation.Buffer.get();
-
-  KAFFINITY coreMask = 0;
-  size_t read = 0;
-  while (read <= logicalProcessorInformation.Length) {
-    switch (processorInformation->Relationship) {
-      case RelationProcessorCore:
-        coreMask |= processorInformation->Processor.GroupMask->Mask;
-        break;
-    }
-
-    read += sizeof(SYSTEM_LOGICAL_PROCESSOR_INFORMATION_EX);
-    processorInformation++;
-  }
-  return __popcnt64(coreMask);
-}
 
 uint32_t LearningModelDevice::NumberOfIntraOpThreads() {
   if (IsCpuDevice()) {
-    return static_cast<uint32_t>(GetNumberOfCores() - GetNumberOfSoCDieAtoms());
+    return HardwareCoreEnumerator::DefaultIntraOpNumThreads();
   } else {
     // GPU sessions should not rely on intra op threads.
     // Creating a large thread pool is unnecessary and wasteful, and can cause
